requirements.txt
discord.py
ffmpeg
PyNaCl